Move reorder cards functionality to thunk
